requirements.txt
google-generativeai>=0.3.0
requests>=2.31.0
pandas>=2.0.0
beautifulsoup4>=4.12.0
openpyxl>=3.1.0
matplotlib>=3.7.0
pillow>=10.0.0
numpy>=1.24.0
nltk>=3.8.0
ollama>=0.1.0
transformers>=4.30.0
torch>=2.0.0
langchain>=0.1.0
chromadb>=0.4.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0